requirements.txt
django==5.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1
drf-yasg==1.21.10
python-dotenv==1.0.0
pyyaml==6.0.2
inflection==0.5.1
packaging==25.0
uritemplate==4.1.1
django-model-utils==5.0.0
django-redis==5.4.0
